+++ apps/client/package.json
@@ -13,7 +13,6 @@
     "@ckeditor/ckeditor5-react": "^6.2.0",
     "@plaiceholder/next": "^3.0.0",
     "aos": "^2.3.4",
-    "bcrypt": "^5.1.1",
     "dayjs": "^1.11.10",
     "firebase": "workspace:*",
     "form-data": "^4.0.0",
@@ -44,6 +43,7 @@
     "@types/uuid": "^9.0.8",
     "autoprefixer": "^10.0.1",
     "axios": "^1.6.7",
+    "bcrypt": "^5.1.1",
     "cookie": "^0.6.0",
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
