+++ packages/helper/package.json
@@ -10,12 +10,10 @@
   "devDependencies": {
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
-    "@types/bcrypt": "^5.0.2",
     "@types/eslint": "^8.44.7",
     "@types/node": "^20.5.2",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
-    "bcrypt": "^5.1.1",
     "eslint": "^8.53.0",
     "immer": "^10.1.1",
     "react": "^18.2.0",
