--- packages/helper/src/bcrypt/comparePassword.ts
@@ -1,6 +1,0 @@
-import { compare } from "bcrypt";
-
-export const comparePassword = async (password: string, hashedPassword: string) => {
-  const isMatch = await compare(password, hashedPassword);
-  return isMatch;
-};
