--- packages/helper/src/bcrypt/hashPassword.ts
@@ -1,6 +1,0 @@
-import { hash } from "bcrypt";
-
-export const hashPassword = async (password: string) => {
-  const hashedPassword = await hash(password, 10);
-  return hashedPassword;
-};
