+++ packages/helper/src/index.ts
@@ -1,3 +1,1 @@
 export { useDateTab } from "./useDateTab";
-export { hashPassword } from "./bcrypt/hashPassword";
-export { comparePassword } from "./bcrypt/comparePassword";
